requirements.txt
graphrag
django-dotenv